Move testutils into storage to avoid circular dep
